requirements.txt
aiogram==3.27.0
aiohttp==3.9.3
aiohttp-socks==0.8.4
aiosqlite==0.20.0
python-dotenv==1.2.2
yandexgpt-python==0.2.3
pydantic==2.12.5
pytest==8.3.5
pytest-asyncio==0.25.3